int8 is translated as string in JS. Changed them to int4 to return as number.

+++ src/modules/orders/infra/typeorm/entities/OrdersProducts.ts
@@ -33,7 +33,7 @@ class OrdersProducts {
   @Column('real')
   price: number;
 
-  @Column('int8')
+  @Column('int4')
   quantity: number;
 
   @CreateDateColumn()
